kernel: rmt: rmt-ps-default: Removed work conservation from RMT
The RMT design assumed that in case it were possible to send a
PDU, the PDU should be sent. That was a wrong assumption since
scheduling policies could need to delay PDUs to ensure abiding by
customers' contracts. With the new design, is for the scheduling
policies to decide if a packet must be sent or must not. The RMT
continues in charge of sending the PDUs.